event posting to discord

+++ api/routers/reminds_router.py
@@ -30,12 +30,14 @@ async def get_current_events(count: int = 10, skip: int = 0, session = Depends(d
     current_events = session.exec(select(CurrentEvent).offset(skip).limit(count)).all()
     for i in range(len(current_events)):
         current_events[i] = current_events[i].model_dump()
+        current_events[i].update({"delete_date": str(current_events[i].get("delete_date"))})
     return JSONResponse(current_events, status_code=status.HTTP_200_OK)
 
 @router.delete("/{current_event_id}", dependencies=[Depends(require_admin)])
 async def delete_current_event(current_event_id: int, session = Depends(database.get_session)):
     current_event = session.exec(select(CurrentEvent).where(CurrentEvent.id == current_event_id)).first()
-    session.delete(current_event)
-    session.commit()
-
-    return JSONResponse(current_event.model_dump(mode="json"), status_code=status.HTTP_200_OK)
+    if current_event:
+        session.delete(current_event)
+        session.commit()
+        return JSONResponse(current_event.model_dump(mode="json"), status_code=status.HTTP_200_OK)
+    return JSONResponse("No model found.", status_code=status.HTTP_404_NOT_FOUND)
